refactor: move user endpoints to router
